+++ daemon/service/firewall/vpncoexistence/vpnCoexistence_windows.go
@@ -1,9 +1,6 @@
 // TODO FIXME: prepend license
 // Copyright (c) 2024 privateLINE, LLC.
 
-//go:build windows
-// +build windows
-
 package vpncoexistence
 
 import (
@@ -15,11 +12,15 @@ import (
 	"strconv"
 	"strings"
 	"syscall"
+	"time"
+	"unsafe"
 
+	mapset "github.com/deckarep/golang-set/v2"
 	"github.com/swapnilsparsh/devsVPN/daemon/service/firewall/winlib"
 	"github.com/swapnilsparsh/devsVPN/daemon/service/platform"
 	"github.com/swapnilsparsh/devsVPN/daemon/shell"
 	"golang.org/x/sys/windows"
+	"golang.org/x/sys/windows/svc"
 	"golang.org/x/sys/windows/svc/mgr"
 )
 
@@ -36,7 +37,70 @@ type OtherVpnInfoParsed struct {
 }
 
 var (
-	FirstWordRE *regexp.Regexp = regexp.MustCompilePOSIX("^[^[:space:]_\\.-]+") // regexp for the 1st word: "^[^[:space:]_\.-]+"
+	ZeroGUID = syscall.GUID{}
+
+	// blacklist of words that can't be brand name candidates
+	// Vlad - don't stop services whose name starts with "Wireguard", because we don't want to kill our own connection, lol
+	invalidServiceNamePrefixes mapset.Set[string] = mapset.NewSet[string]("microsoft", "windefend", "edge", "intel", "wireguard")
+
+	// Windows service names to try always:
+
+	// SurfShark uses random GUIDs for its sublayer 0xFFFF (named "WireGuard filters") and provider "WireGuard provider", so try its named services always
+	// "Surfshark Service", "Surfshark WireGuard"
+
+	// Must keep both lists in sync. otherVpnDefaultServiceNamePrefixesToTry must contain literal words, not regular expressions.
+	otherVpnDefaultServiceNamePrefixesToTry mapset.Set[string] = mapset.NewSet[string]("mullvad", "expressvpn", "surfshark", "nord", "proton", "ivpn", "hideme",
+		"mozillavpn", "windscribe", "ipvcallout", "ipvan", "cyberghost", "proton", "tunnelbear", "vypervpn", "vyprvpn", "turbovpn", "cloudflarewarp",
+		"urbanvpn", "eddieelevation", "hshld", "hotspotshield", "privateinternetaccess", "hola", "avg", "securevpn")
+	// (?i) for case-insensitive matching, regex matching group not closed in this string - must be closed where used
+	otherVpnDefaultServiceNameBrandsRegexStart = "(?i)^(mullvad|expressvpn|surfshark|nord|proton|ivpn|hideme|mozillavpn|windscribe|ipv(callout|an)|" +
+		"cyberghost|proton|tunnelbear|vype?rvpn|turbovpn|cloudflarewarp|urbanvpn|eddie.?elevation|hshld|hotspotshield|privateinternetaccess|" +
+		"hola(.?vpn|[^a-z0-9]|$)|.*updater_.*_hola|avg|securevpn"
+	otherVpnDefaultServiceNamePrefixesRE = regexp.MustCompile(otherVpnDefaultServiceNameBrandsRegexStart + ")")
+
+	// other VPNs known to us
+
+	// Mullvad
+	mullvadSublayerKey = syscall.GUID{Data1: 0xC78056FF, Data2: 0x2BC1, Data3: 0x4211, Data4: [8]byte{0xAA, 0xDD, 0x7F, 0x35, 0x8D, 0xEF, 0x20, 0x2D}}
+	mullvadProfile     = OtherVpnInfo{
+		name:       "Mullvad VPN",
+		namePrefix: "mullvad",
+		cliPath:    "ProgramFiles/Mullvad VPN/resources/mullvad.exe",
+		cliCmds: otherVpnCliCmds{
+			cmdStatus:                          "status",
+			statusConnectedRE:                  "^Connected([^a-zA-Z0-9]|$)", // must be 1st line
+			statusDisconnectedRE:               "^Disconnected([^a-zA-Z0-9]|$)",
+			cmdEnableSplitTun:                  []string{"split-tunnel", "set", "on"},
+			cmdAddOurBinaryToSplitTunWhitelist: []string{"split-tunnel", "app", "add"},
+			cmdConnect:                         "connect",
+			cmdDisconnect:                      "disconnect",
+		},
+	}
+
+	// TODO: NordVPN CLI
+	// NordVPN
+	nordVpnSublayerKey = syscall.GUID{Data1: 0x92C759E5, Data2: 0x03BA, Data3: 0x41AD, Data4: [8]byte{0xA5, 0x99, 0x68, 0xAF, 0x2C, 0x1A, 0x17, 0xE5}}
+	nordVpnProfile     = OtherVpnInfo{
+		name:       "NordVPN",
+		namePrefix: "nord",
+		cliPath:    "",
+		//cliPath:    "ProgramFiles/NordVPN/NordVPN.exe", // disabled for now, since we didn't find yet a programmatic way to check whether VPN is connected
+		cliCmds: otherVpnCliCmds{
+			cmdStatus:                          "",
+			statusConnectedRE:                  "",
+			statusDisconnectedRE:               "",
+			cmdEnableSplitTun:                  []string{},
+			cmdAddOurBinaryToSplitTunWhitelist: []string{},
+			cmdConnect:                         "--connect",
+			cmdDisconnect:                      "--disconnect",
+		},
+	}
+
+	// index (DB) of other VPNs by sublayer key
+	otherVpnsBySublayerGUID = map[syscall.GUID]*OtherVpnInfo{
+		mullvadSublayerKey: &mullvadProfile,
+		nordVpnSublayerKey: &nordVpnProfile,
+	}
 )
 
 func (otherVpn *OtherVpnInfoParsed) Close() {
@@ -112,7 +176,7 @@ func lookupOtherVpnProvider(otherVpnProviderKey syscall.GUID, manager *winlib.Ma
 // (1) Try to find the other VPN in our DB via sublayer GUID, or sublayer name, or provider name. Also try to make use of provider.serviceName.
 // (2) Else look for Windows services whose names start with name prefix of any VPN known to us. Also try to make use of provider.serviceName.
 // (3) Else try the list of default service name prefixes, contanining common VPN name brand names.
-func ParseOtherVpn(otherSublayerFound bool, otherSublayer *winlib.SubLayer, manager *winlib.Manager) (otherVpn *OtherVpnInfoParsed /*, err error*/) {
+func ParseOtherVpnBySublayerGUID(otherSublayerFound bool, otherSublayer *winlib.SubLayer, manager *winlib.Manager) (otherVpn *OtherVpnInfoParsed /*, err error*/) {
 	var (
 		err                                      error
 		otherVpnReadonly                         *OtherVpnInfo
@@ -305,4 +369,176 @@ func (otherVpn *OtherVpnInfoParsed) PostSteps() {
 	}
 
 	log.Debug("PostSteps() ending for " + otherVpn.name)
+}
+
+// ================================ Windows service helpers: ================================
+
+// Refer to these code samples:
+//	https://learn.microsoft.com/en-us/windows/win32/services/stopping-a-service
+//	https://learn.microsoft.com/en-us/windows/win32/services/starting-a-service
+//	https://github.com/shirou/gopsutil/blob/master/winservices
+//	https://opensource.srlabs.de/projects/srl_gobuster/repository/11/revisions/master/annotate/src/golang.org/x/sys/windows/svc/example/manage.go
+
+type scmanager struct {
+	mgr *mgr.Mgr
+}
+
+func OpenSCManager() (*scmanager, error) {
+	m, err := mgr.Connect()
+	if err != nil {
+		return nil, err
+	}
+	return &scmanager{m}, nil
+}
+
+func (sc *scmanager) Close() error {
+	return sc.mgr.Disconnect()
+}
+
+// OpenServiceIfRunning - returns opened service if its running. Returns (nil, nil), if service is not running, but no errors.
+func (sc *scmanager) OpenServiceIfRunning(serviceName string) (service *mgr.Service, err error) {
+	if service, err = sc.mgr.OpenService(serviceName); err != nil {
+		return nil, err
+	}
+
+	// check whether other VPN service was running
+	var serviceStatus ServiceStatus
+	if serviceStatus, err = QueryServiceStatusEx(service); err != nil {
+		service.Close()
+		return nil, log.ErrorE(fmt.Errorf("error QueryServiceStatusEx(): %w", err), 0)
+	}
+	if serviceStatus.State != svc.Running {
+		service.Close()
+		return nil, nil
+	}
+
+	return service, nil
+}
+
+// lowercaseSvcNameFromProvider can be ""
+// otherVpnSvcNameRE must be configured for case-insensitive matching
+func (sc *scmanager) FindMatchingRunningServices(serviceNameFromProvider string, otherVpnSvcNameRE *regexp.Regexp) (servicesRunning []*mgr.Service, err error) {
+	servicesRunning = make([]*mgr.Service, 0, 1)
+
+	var svcList []string
+	if svcList, err = sc.mgr.ListServices(); err != nil {
+		return servicesRunning, err
+	}
+
+	lowercaseSvcNameFromProvider := strings.ToLower(serviceNameFromProvider)
+	log.Debug("FindMatchingRunningServices(): lowercaseSvcNameFromProvider='" + lowercaseSvcNameFromProvider + "' otherVpnSvcNameRE='" + otherVpnSvcNameRE.String() + "'")
+	for _, svcName := range svcList {
+		if otherVpnSvcNameRE.MatchString(svcName) || strings.ToLower(svcName) == lowercaseSvcNameFromProvider {
+			if svc, err := sc.OpenServiceIfRunning(svcName); err != nil {
+				log.ErrorFE("error opening service '%s': %w", svcName, err)
+			} else if svc != nil {
+				servicesRunning = append(servicesRunning, svc)
+				// if len(servicesRunning) >= MAX_WINDOWS_SERVICE_CANDIDATES {
+				// 	return servicesRunning, nil
+				// }
+			}
+		}
+	}
+
+	return servicesRunning, nil
+}
+
+// ServiceStatus combines State and Accepted commands to fully describe running service.
+type ServiceStatus struct {
+	State         svc.State
+	Accepts       svc.Accepted
+	Pid           uint32
+	Win32ExitCode uint32
+}
+
+// QueryServiceStatusEx return the specified name service currentState and ControlsAccepted
+func QueryServiceStatusEx(service *mgr.Service) (ServiceStatus, error) {
+	var p *windows.SERVICE_STATUS_PROCESS
+	var bytesNeeded uint32
+	var buf []byte
+
+	if err := windows.QueryServiceStatusEx(service.Handle, windows.SC_STATUS_PROCESS_INFO, nil, 0, &bytesNeeded); err != windows.ERROR_INSUFFICIENT_BUFFER {
+		return ServiceStatus{}, err
+	}
+
+	buf = make([]byte, bytesNeeded)
+	p = (*windows.SERVICE_STATUS_PROCESS)(unsafe.Pointer(&buf[0]))
+	if err := windows.QueryServiceStatusEx(service.Handle, windows.SC_STATUS_PROCESS_INFO, &buf[0], uint32(len(buf)), &bytesNeeded); err != nil {
+		return ServiceStatus{}, err
+	}
+
+	return ServiceStatus{
+		State:         svc.State(p.CurrentState),
+		Accepts:       svc.Accepted(p.ControlsAccepted),
+		Pid:           p.ProcessId,
+		Win32ExitCode: p.Win32ExitCode,
+	}, nil
+}
+
+// Making a copy of Control(), because we need waitHint for this particular command
+// func controlServiceExt(s *mgr.Service, c svc.Cmd) (svc.Status, error) {
+// 	var t windows.SERVICE_STATUS
+// 	err := windows.ControlService(s.Handle, uint32(c), &t)
+// 	if err != nil &&
+// 		err != windows.ERROR_INVALID_SERVICE_CONTROL &&
+// 		err != windows.ERROR_SERVICE_CANNOT_ACCEPT_CTRL &&
+// 		err != windows.ERROR_SERVICE_NOT_ACTIVE {
+// 		return svc.Status{}, err
+// 	}
+// 	return svc.Status{
+// 		State:    svc.State(t.CurrentState),
+// 		Accepts:  svc.Accepted(t.ControlsAccepted),
+// 		WaitHint: t.WaitHint,
+// 	}, err
+// }
+
+func controlService(s *mgr.Service, c svc.Cmd, to svc.State) error {
+	status, err := s.Control(c)
+	if err != nil {
+		return fmt.Errorf("could not send control=%d: %v", c, err)
+	}
+
+	timeout := time.Now().Add(MAX_WAIT)
+	for status.State != to {
+		if timeout.Before(time.Now()) {
+			return fmt.Errorf("timeout waiting for service to go to state=%d", to)
+		}
+		time.Sleep(300 * time.Millisecond)
+		status, err = s.Query()
+		if err != nil {
+			return fmt.Errorf("could not retrieve service status: %v", err)
+		}
+	}
+	return nil
+}
+
+func StopService(s *mgr.Service) error {
+	serviceStatus, err := QueryServiceStatusEx(s)
+	// log.Debug("service '" + s.Name + "' state queried")
+	if err != nil {
+		return log.ErrorE(fmt.Errorf("error QueryServiceStatusEx(): %w", err), 0)
+	}
+
+	if serviceStatus.State == svc.Stopped {
+		log.Debug(fmt.Sprintf("service '%s' already stopped", s.Name))
+		return nil
+	}
+
+	// TODO wait for StopPending?
+
+	return controlService(s, svc.Stop, svc.Stopped)
+}
+
+func StartService(s *mgr.Service) error {
+	serviceStatus, err := QueryServiceStatusEx(s)
+	if err != nil {
+		return log.ErrorE(fmt.Errorf("error QueryServiceStatusEx(): %w", err), 0)
+	}
+
+	if serviceStatus.State == svc.Running {
+		log.Debug(fmt.Sprintf("service '%s' already running", s.Name))
+		return nil
+	}
+
+	return s.Start()
 }
